You need more comments to explain.
Even though this isn't Solidity development you should consider using the NatSpec format for the functions. Even for Java it can be useful because it looks so much more clean when specifying the use, parameters and titles of a public function, method, ect.

Here is the article about NatSpec here: https://docs.soliditylang.org/en/develop/natspec-format.html

Also, I threw together the params just for a vague reference. It may not be right in the 10 seconds I spent writing it.